chore: refactor giant pox-2 test file into multiple files, run separately in ci
