+++ tests/jest.config.2.1.js
@@ -1,9 +1,15 @@
-module.exports = {
+/** @type {import('jest').Config} */
+const config = {
   preset: 'ts-jest',
   testEnvironment: 'node',
   rootDir: `${require('path').dirname(__dirname)}/src`,
   testMatch: ['<rootDir>/tests-2.1/**/*.ts'],
-  testPathIgnorePatterns: ['<rootDir>/tests-2.1/global-setup.ts', '<rootDir>/tests-2.1/global-teardown.ts', '<rootDir>/tests-2.1/env-setup.ts'],
+  testPathIgnorePatterns: [
+    '<rootDir>/tests-2.1/global-setup.ts', 
+    '<rootDir>/tests-2.1/global-teardown.ts', 
+    '<rootDir>/tests-2.1/env-setup.ts',
+    '<rootDir>/tests-2.1/test-helpers.ts',
+  ],
   collectCoverageFrom: ['<rootDir>/**/*.ts'],
   coveragePathIgnorePatterns: ['<rootDir>/tests*'],
   coverageDirectory: '<rootDir>/../coverage',
@@ -12,4 +18,7 @@ module.exports = {
   setupFilesAfterEnv: ['<rootDir>/tests-2.1/env-setup.ts'],
   testTimeout: 120000,
   verbose: true,
-}
+  bail: true,
+};
+
+module.exports = config;
